성별,활동량 출력 추가

+++ src/main/java/com/example/demo/metabolism/domain/Metabolism.java
@@ -10,7 +10,7 @@ public class Metabolism {
     private int height;
     private int weight;
     private String gender;
-    private int activity;
+    private double activity;
     private int basalMetabolicRate;
     private int activityMetabolism;
 }
